Convert JWT-Knox into a ViewSet (#7)
Refactor JWT-KNox into a ViewSet

In this commit we also move Auth/JWT URLs to be in the API app.

xea_core already includes the API Django app to include the ^/api/
routes, it makes no sense to have another ^/api/xxx route outside that
app for the moment. Those configurations should be imported by the api
app.

Now the ViewSet needs a different authentication_classes depending on
whether we are using get_token or the rest of the ViewSet functions.
Because of that, we need to implement get_authenticators.

Instead of implementing get_authenticators manually, as it will be
called by super().initialize_request(request, *args, **kwargs), we
implement a PerViewAuthentication mixin instead, that, hopefully, we
will be able to reuse elsewhere.

This way, we re-call initialize_request in initial(), after the
ViewSet's request has already been initialized and thus we are able to
use self.action in the subclassed per-view authenticator logic (and
users of the mixin needn't be aware on the ViewSet initialization
lifecycle).

URLs in jwt-knox are are no longer inside a namespace, because the
router already prefixes urls with jwt_knox, so it's feels bit redundant.

JWT-Knox tests are modified to use .force_authenticate instead of .login

Client.login() is a Django Client function which leverages
SessionAuthentication, and its lifecycle seems not to be compatible with
our custom mixin's patching. So, instead of using SessionAuthentication
we can either use BasicAuthentication (which seems not to suffer from
this issue) or using rest_framework's APIClient.force_authenticate,
which understands rest_framework's Request object and forces
authentication on it, instead of relying on a Django Session.

This change forces us to cache self.user (the created user model) to
provide it to the request.